% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeans.R
\docType{data}
\name{jeans}
\alias{jeans}
\title{Characteristics of a selection of jeans on Anthropologie's website.}
\format{
A tibble where each row represents a pair of jeans.  There are 4 rows and 10 variables:
\describe{
\item{brand}{a character, brand name of jeans}
\item{name}{a character, name of jeans}
\item{cost}{a numeric, cost (before shipping) in USD}
\item{style}{a character, style of jeans}
\item{wash}{a character, wash/color of jeans}
\item{sizes_available}{a character, all available sizes on 9/12/2025}
\item{made_in_USA}{a character for whether or not the jeans were made in the USA}
\item{rise}{a numeric, rise of jeans in inches}
\item{inseam}{a numeric, inseam of jeans in inches}
\item{leg_opening}{a numeric, leg opening of jeans in inches}
}
}
\source{
{anthropologie.com}
}
\usage{
jeans
}
\description{
Data were collected on 9/12/2025 and include variables related to the fit, style, brand, and cost of the jeans.
}
\keyword{datasets}
